requirements.txt
asgiref==3.3.1
Django==3.1.7
django-countries==7.1
django-crispy-forms==1.11.1
django-tinymce==3.2.0
Pillow==8.1.2
pytz==2021.1
pyuca==1.2
sqlparse==0.4.1
wfastcgi==3.0.0
